requirements.txt
openai
pandas
fastapi
uvicorn
jinja2
python-multipart
mysql-connector-python
requests
